Introduce serverless host (WIP)
